refactor: rename types
