+++ apps/web/src/services/dashboards/shared/dashboard-widget-input-form/helpers/schema-helper.ts
@@ -135,3 +135,4 @@ export const getRefinedWidgetInheritOptions = (widgetInfo: DashboardLayoutWidget
     }
     return inheritOptions;
 };
+
